ViewController에서 CollectionView 분리, CollectionCell Size 100x
