+++ PhotoLibraryApp/PhotoLibraryApp/ViewController.swift
@@ -9,35 +9,9 @@ import UIKit
 
 class ViewController: UIViewController {
     
-    @IBOutlet weak var photoCollectionView: UICollectionView!
+    @IBOutlet weak var photoCollectionView: PhotoCollectionView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        configureCollectionView()
-    }
-    
-    func configureCollectionView() {
-        photoCollectionView.dataSource = self
-        photoCollectionView.delegate = self
     }
 }
-
-extension ViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 40
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "PhotoCell", for: indexPath) as? PhotoCell else {
-            return UICollectionViewCell()
-        }
-        return cell
-    }
-}
-
-extension ViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 80, height: 80)
-    }
-}
-
